+++ tests/test_context.py
@@ -18,86 +18,28 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-# from __future__ import absolute_import
-#
-# import threading
-#
-# import pytest
-# from basictracer import BasicTracer, SpanRecorder
-# from tchannel import Response
-# from tchannel import TChannel
-# from tchannel import schemes
-# from tornado import gen
-#
-#
-# @pytest.fixture
-# def span_recorder():
-#     class InMemoryRecorder(SpanRecorder):
-#         def __init__(self):
-#             self.spans = []
-#             self.mux = threading.Lock()
-#
-#         def record_span(self, span):
-#             with self.mux:
-#                 self.spans.append(span)
-#
-#         def get_spans(self):
-#             with self.mux:
-#                 return self.spans[:]
-#
-#     return InMemoryRecorder()
-#
-#
-# # noinspection PyShadowingNames
-# @pytest.fixture
-# def tracer(span_recorder):
-#     return BasicTracer(recorder=span_recorder)
-#
-#
-# @pytest.mark.gen_test
-# @pytest.mark.call
-# def test_context_should_carry_tracing_info(tracer):
-#
-#     context = [None, None]
-#     server = TChannel(name='server', trace=tracer)
-#
-#     @server.register(scheme=schemes.RAW)
-#     @gen.coroutine
-#     def endpoint1(request):
-#         yield server.call(  # make recursive call to the 2nd endpoint
-#             scheme=schemes.RAW,
-#             service='server',
-#             arg1='endpoint2',
-#             arg2='req headers',
-#             arg3='req body',
-#             hostport=server.hostport,
-#         )
-#         context[0] = server.context_provider.get_current_span()
-#         raise gen.Return(Response('resp body', 'resp headers'))
-#
-#     @server.register(scheme=schemes.RAW)
-#     def endpoint2(request):
-#         context[1] = server.context_provider.get_current_span()
-#         return Response('resp body', 'resp headers')
-#
-#     server.listen()
-#
-#     # Make a call:
-#
-#     tchannel = TChannel(name='client', trace=tracer)
-#
-#     span = tracer.start_span('root-span')
-#     span.set_baggage_item('bender', 'is great')
-#     with tchannel.context_provider.span_in_context(span):
-#         f = tchannel.call(
-#             scheme=schemes.RAW,
-#             service='server',
-#             arg1='endpoint1',
-#             arg2='req headers',
-#             arg3='req body',
-#             hostport=server.hostport,
-#         )
-#     yield f
-#
-#     assert context[0].get_baggage_item('bender') == 'is great'
-#     assert context[1].get_baggage_item('bender') == 'is great'
+from __future__ import absolute_import
+
+import pytest
+from tchannel import context
+from tornado import gen
+
+
+@pytest.mark.gen_test
+def test_deprecated_context_provider():
+    """
+    Test that the deprecated RequestContextProvider() can still propagate
+    the request context across coroutines.
+    """
+    provider = context.RequestContextProvider()
+
+    @gen.coroutine
+    def _get_context():
+        res = provider.get_current_context().parent_tracing
+        raise gen.Return(res)
+
+    with provider.request_context(parent_tracing='Farnsworth'):
+        res = _get_context()
+
+    res = yield res
+    assert res == 'Farnsworth'