+++ tests/test_tracing.py
@@ -33,8 +33,8 @@ import tornado
 import tornado.gen
 import tornado.testing
 import tornado.web
-from basictracer import BasicTracer
-from basictracer.recorder import InMemoryRecorder
+from jaeger_client import Tracer, ConstSampler
+from jaeger_client.reporter import InMemoryReporter
 from opentracing import Format
 from opentracing.ext import tags
 from opentracing_instrumentation.client_hooks.tornado_http import (
@@ -69,9 +69,26 @@ tornado.testing.bind_unused_port = patched_bind_unused_port
 
 
 # noinspection PyShadowingNames
-@pytest.fixture
+@pytest.yield_fixture
 def tracer():
-    return BasicTracer(recorder=InMemoryRecorder())
+    reporter = InMemoryReporter()
+    report_func = reporter.report_span
+
+    def log_and_report(span):
+        print 'Reporting span %s' % span
+        report_func(span)
+
+    reporter.report_span = log_and_report
+
+    tracer = Tracer(
+        service_name='test-tracer',
+        sampler=ConstSampler(True),
+        reporter=reporter,
+    )
+    try:
+        yield tracer
+    finally:
+        tracer.close()
 
 
 @pytest.yield_fixture
@@ -92,6 +109,7 @@ def thrift_service(tmpdir):
           string thrift1(1: string hostport), // calls thrift2()
           string thrift2(),
           string thrift3(1: string hostport)  // calls http
+          string thrift4(1: string hostport)  // calls raw
         }
     ''')
 
@@ -132,7 +150,7 @@ def register(tchannel, thrift_service, http_client, base_url):
     def _extract_span():
         span = get_current_span()
         if span:
-            return span.context.get_baggage_item(BAGGAGE_KEY)
+            return span.get_baggage_item(BAGGAGE_KEY)
         return "baggage not found"
 
     @tchannel.json.register('endpoint3')
@@ -151,6 +169,24 @@ def register(tchannel, thrift_service, http_client, base_url):
                 response.code, response.body
             ))
         raise tornado.gen.Return(Response(body=response.body))
+
+    @tchannel.thrift.register(thrift_service.X, method='thrift4')
+    @tornado.gen.coroutine
+    def thrift4(request):
+        print 'thrift4 hit'
+        host_port = request.body.hostport
+        res = tchannel.raw(
+            service='test-service',
+            endpoint='raw2',
+            hostport=host_port,
+        )
+        res = yield res
+        print 'result', res
+        raise tornado.gen.Return(Response(res.body))
+
+    @tchannel.raw.register('raw2')
+    def raw2(_):
+        return _extract_span()
 
 
 @pytest.fixture
@@ -191,7 +227,7 @@ class HttpHandler(tornado.web.RequestHandler):
     def get(self):
         span = self._get_span()
         if span:
-            self.write(span.context.get_baggage_item(BAGGAGE_KEY))
+            self.write(span.get_baggage_item(BAGGAGE_KEY))
             self.set_status(200)
             span.finish()
 
@@ -220,26 +256,30 @@ class HttpHandler(tornado.web.RequestHandler):
                 span.finish()
 
 
-@pytest.mark.parametrize('endpoint,transport,encoding,enabled,expect_spans', [
-    # tchannel(json)->tchannel(json)
-    ('endpoint1', 'tchannel', 'json',   True,  5),
-    ('endpoint1', 'tchannel', 'json',   False, 0),
-    # tchannel(thrift)->tchannel(thrift)
-    ('thrift1',   'tchannel', 'thrift', True,  5),
-    ('thrift1',   'tchannel', 'thrift', False, 0),
-    # tchannel(json)->http(json)
-    ('endpoint3', 'tchannel', 'json',   True,  5),
-    ('endpoint3', 'tchannel', 'json',   False, 0),
-    # tchannel(thrift)->http(json)
-    ('thrift3',   'tchannel', 'thrift', True,  5),
-    ('thrift3',   'tchannel', 'thrift', False, 0),
-    # http->tchannel
-    ('/',         'http',     'json',   True,  5),
-    ('/',         'http',     'json',   False, 0),
-])
+@pytest.mark.parametrize(
+    'endpoint,transport,encoding,enabled,expect_spans,expect_baggage', [
+        # tchannel(json)->tchannel(json)
+        ('endpoint1', 'tchannel', 'json', True, 5, True),
+        ('endpoint1', 'tchannel', 'json', False, 0, True),
+        # tchannel(thrift)->tchannel(thrift)
+        ('thrift1', 'tchannel', 'thrift', True, 5, True),
+        ('thrift1', 'tchannel', 'thrift', False, 0, True),
+        # tchannel(thrift)->tchannel(raw)
+        ('thrift4', 'tchannel', 'thrift', True, 5, False),
+        ('thrift4', 'tchannel', 'thrift', False, 0, False),
+        # tchannel(json)->http(json)
+        ('endpoint3', 'tchannel', 'json', True, 5, True),
+        ('endpoint3', 'tchannel', 'json', False, 0, True),
+        # tchannel(thrift)->http(json)
+        ('thrift3', 'tchannel', 'thrift', True, 5, True),
+        ('thrift3', 'tchannel', 'thrift', False, 0, True),
+        # http->tchannel
+        ('/', 'http', 'json', True, 5, True),
+        ('/', 'http', 'json', False, 0, True),
+    ])
 @pytest.mark.gen_test
 def test_trace_propagation(
-        endpoint, transport, encoding, enabled, expect_spans,
+        endpoint, transport, encoding, enabled, expect_spans, expect_baggage,
         http_patchers, tracer, mock_server, thrift_service,
         app, http_server, base_url, http_client):
     """
@@ -300,7 +340,7 @@ def test_trace_propagation(
 
         span = tracer.start_span('root')
         baggage = 'from handler3 %d' % time.time()
-        span.context.set_baggage_item(BAGGAGE_KEY, baggage)
+        span.set_baggage_item(BAGGAGE_KEY, baggage)
         if not enabled:
             span.set_tag('sampling.priority', 0)
         with span:  # use span as context manager so that it's always finished
@@ -315,10 +355,23 @@ def test_trace_propagation(
                             body=mock_server.hostport,
                         )
                     elif encoding == 'thrift':
-                        response_future = tchannel.thrift(
-                            thrift_service.X.thrift1(mock_server.hostport),
-                            hostport=mock_server.hostport,
-                        )
+                        if endpoint == 'thrift1':
+                            response_future = tchannel.thrift(
+                                thrift_service.X.thrift1(mock_server.hostport),
+                                hostport=mock_server.hostport,
+                            )
+                        elif endpoint == 'thrift3':
+                            response_future = tchannel.thrift(
+                                thrift_service.X.thrift3(mock_server.hostport),
+                                hostport=mock_server.hostport,
+                            )
+                        elif endpoint == 'thrift4':
+                            response_future = tchannel.thrift(
+                                thrift_service.X.thrift4(mock_server.hostport),
+                                hostport=mock_server.hostport,
+                            )
+                        else:
+                            raise ValueError('wrong endpoint %s' % endpoint)
                     else:
                         raise ValueError('wrong encoding %s' % encoding)
                 elif transport == 'http':
@@ -335,10 +388,11 @@ def test_trace_propagation(
             response = yield response_future
 
     body = response.body
-    assert body == baggage
+    if expect_baggage:
+        assert body == baggage
 
     def get_sampled_spans():
-        return [s for s in tracer.recorder.get_spans() if s.context.sampled]
+        return [s for s in tracer.reporter.get_spans() if s.is_sampled]
 
     # Sometimes the test runs into weird race condition where the
     # after_send_response() hook is executed, but the span is not yet
@@ -353,88 +407,10 @@ def test_trace_propagation(
     spans = get_sampled_spans()
     assert expect_spans == len(spans), 'Unexpected number of spans reported'
     # We expect all trace IDs in collected spans to be the same
-    spans = tracer.recorder.get_spans()
-    assert 1 == len(set([s.context.trace_id for s in spans])), \
-        'all spans must have the same trace_id'
-
-
-# @pytest.mark.parametrize('ex_cls,start_span', [
-#     (opentracing.UnsupportedFormatException, 0),
-#     (opentracing.InvalidCarrierException, 0),
-#     (opentracing.TraceCorruptedException, 1),
-# ])
-# def test_before_receive_request(ex_cls, start_span):
-#     class Tracer(opentracing.Tracer):
-#         def __init__(self):
-#             super(Tracer, self).__init__()
-#             self.span_started = 0
-#
-#         def join(self, operation_name, format, carrier):
-#             raise ex_cls()
-#
-#         def start_span(self, operation_name=None,
-#                        parent=None, tags=None, start_time=None):
-#             self.span_started += 1
-#             return super(Tracer, self).start_span(
-#                 operation_name=operation_name)
-#
-#     log_fn = mock.Mock()
-#     hook = OpenTracingHook(context_provider=None, tracer=Tracer(),
-#                            log_exception_fn=log_fn)
-#     hook.before_receive_request(request=mock.Mock())
-#     assert log_fn.call_count == 1
-#     assert hook.tracer.span_started == start_span
-#
-#
-# @pytest.mark.gen_test
-# def test_span_to_trace(tracer, mock_server):
-#     """
-#     In this test we verify that if the tracer implementation supports the
-#     notions of trace_id, span_id, parent_id (similar to Zipkin) then we can
-#     pass those IDs not just via the headers (primary OpenTracing propagation
-#     mechanism), but also via TChannel's built-in tracing slot in the frame.
-#     :param tracer: injected BasicTracer mixin
-#     :param mock_server: injected TChannel mock server
-#     """
-#     def span_to_trace(span):
-#         return {
-#             'trace_id': span.context.trace_id,
-#             'span_id': span.context.span_id,
-#             'parent_span_id': span.context.parent_id,
-#         }
-#
-#     context_provider = OpenTracingRequestContextProvider()
-#     hook = OpenTracingHook(tracer=tracer, context_provider=context_provider,
-#                            span_to_trace_fn=span_to_trace)
-#
-#     @mock_server.tchannel.raw.register('endpoint1')
-#     @tornado.gen.coroutine
-#     def handler1(request):
-#         ctx = mock_server.tchannel.context_provider.get_current_context()
-#         if hasattr(ctx, 'parent_tracing'):
-#             res = ctx.parent_tracing.trace_id
-#         else:
-#             res = 'unknown'
-#         raise tornado.gen.Return(Response('%s' % res))
-#
-#     tchannel = TChannel(name='test', context_provider=context_provider)
-#     tchannel.hooks.register(hook)
-#
-#     with mock.patch('opentracing.tracer', tracer):
-#         assert opentracing.tracer == tracer  # sanity check that patch worked
-#         span = tracer.start_span('root')
-#        with span:  # use span as context manager so that it's always finished
-#             wrapper = SpanWrapper(span=span)
-#             with context_provider.request_context(wrapper):
-#                 response_future = tchannel.raw(
-#                     service='test-client',
-#                     hostport=mock_server.hostport,
-#                     endpoint='endpoint1',
-#                     headers=mock_server.hostport,
-#                     trace=False,
-#                 )
-#             response = yield response_future
-#         assert span.context.trace_id == long(response.body)
+    if expect_spans > 0:
+        spans = tracer.reporter.get_spans()
+        assert 1 == len(set([s.trace_id for s in spans])), \
+            'all spans must have the same trace_id'
 
 
 @pytest.mark.parametrize('encoding,operation', [
@@ -448,7 +424,7 @@ def test_span_tags(encoding, operation, tracer, thrift_service):
 
     def get_span_baggage():
         sp = server.context_provider.get_current_span()
-        baggage = sp.context.get_baggage_item('bender') if sp else None
+        baggage = sp.get_baggage_item('bender') if sp else None
         return {'bender': baggage}
 
     @server.json.register('foo')
@@ -459,10 +435,10 @@ def test_span_tags(encoding, operation, tracer, thrift_service):
     def thrift2(_):
         return json.dumps(get_span_baggage())
 
-    client = TChannel('client', tracer=tracer)
+    client = TChannel('client', tracer=tracer, trace=True)
 
     span = tracer.start_span('root')
-    span.context.set_baggage_item('bender', 'is great')
+    span.set_baggage_item('bender', 'is great')
     with span:
         res = None
         with client.context_provider.span_in_context(span):
@@ -485,26 +461,34 @@ def test_span_tags(encoding, operation, tracer, thrift_service):
     if isinstance(res, basestring):
         res = json.loads(res)
     assert res == {'bender': 'is great'}
-    spans = tracer.recorder.get_spans()
+    for i in range(1000):
+        spans = tracer.reporter.get_spans()
+        if len(spans) == 3:
+            break
+        yield tornado.gen.sleep(0.001)  # yield execution and sleep for 1ms
+    spans = tracer.reporter.get_spans()
     assert len(spans) == 3
-    assert 1 == len(set([s.context.trace_id for s in spans])), \
+    assert 1 == len(set([s.trace_id for s in spans])), \
         'all spans must have the same trace_id'
     parent = child = None
     for s in spans:
         if s.tags is None:
             continue
-        if s.tags.get(tags.SPAN_KIND, None) == tags.SPAN_KIND_RPC_SERVER:
+        print('tags %s' % s.tags)
+        # replace list with dictionary
+        s.tags = {tag.key: tag.value for tag in s.tags}
+        if s.kind == tags.SPAN_KIND_RPC_SERVER:
             child = s
-        elif s.tags.get(tags.SPAN_KIND, None) == tags.SPAN_KIND_RPC_CLIENT:
+        elif s.kind == tags.SPAN_KIND_RPC_CLIENT:
             parent = s
     assert parent is not None
     assert child is not None
     assert parent.operation_name == operation
     assert child.operation_name == operation
-    assert parent.tags.get(tags.PEER_SERVICE) == 'test-service'
-    assert child.tags.get(tags.PEER_SERVICE) == 'client'
-    assert parent.tags.get(tags.PEER_HOST_IPV4) is not None
-    assert child.tags.get(tags.PEER_HOST_IPV4) is not None
+    assert parent.peer['service_name'] == 'test-service'
+    assert child.peer['service_name'] == 'client'
+    assert parent.peer['ipv4'] is not None
+    assert child.peer['ipv4'] is not None
     assert parent.tags.get('as') == encoding
     assert child.tags.get('as') == encoding